Reorg tests
Split the tests into separate task files so that they can be imported by other
testing tools.
Currently, it is planned that the collectd_config[1] and tripleo_collectd[2]
roles will use this in molecule and CI testing.
Later on, it'll be used to test other components intended for use with tripleo.

[1] https://github.com/infrawatch/collectd-config-ansible-role
[2] https://github.com/infrawatch/tripleo-collectd-ansible-role