Remove the docker->podman symlink
This symlink was being used as a workaround for OSP16 and OSP13 using
podman and docker, respectively, so that two branchs/sets of tests
wouldn't have to be maintained.

This was not a good solution as it left the symlink in place, and
blocks upgrades on clouds where this has been run, since podman
already exists at the time of upgrade.

Instead of the symlink, and selecting docker vs podman based on the
OSP13 tag being passed, a ``container_bin`` var is set based on
whether podman exists on the system or not.
``container_bin`` default to ``podman`` and is updated to
``docker`` if ``/usr/bin/podman`` doesn't exist.