[collectd] Update test for metrics collected
Some metrics produce more than one value at a time (e.g. rx and tx
values for network/if_octets)
The test will fail when the number of values returned is not one.
This change updates that test to make sure that there is a non-zero
number of values returned.