Increasing the loop size for snmpTraps test
The default alertmanager route we describe in the docs requires 30s of disruption at minimum before it will send an alert.
We don't want to sleep for long because the pod may be able to restart and transmit some data,
so we should keep the sleep 1 but increase the loop size to guarantee that this lasts at least 30s.