[ci] Run report_result in a separate hook
Report result should gather the results files and make them available
on the controller.
If the callback plugin that is being used aggregates results at the end
of the play, the this file is not available in the test playbook.

Use copy with remote src instead of shell/cp, so that the files can be
copied from compute nodes into the controller logs dir.
Add a new play to make sure the output dir is created on the controller
before copying the files.