Use ignore_errors in plays for test_logging
The logging plugins we use need all the tests to run in order to report properly.
Adding ignore_errors to the plays that run the tests allows all the tests to run even when one fails.

Also added verbosity to the ansible config to get more info about the failing tests