Determine API versions based on observability startegy
Depending on the observability strategy that STF uses, observability
components will be installed from community or Red hat sources. These
crate the same CRDs, but with different APIs e.g.
prometheusrules.monitoring.rhobs vs
prometheusrules.monitoring.coreos.com. An apiVersion needs to be set
when creating these resources during the tests, and the tests will fail
if the correct version is not available.

The test_alerts role will select which observability APIs to use based
on which observability_straegy is set. The tests will also explicitly
use the fully qualified name of the objects using ``oc get``.

The following roles are updated:
* test_alerts
* test_verify_email