Rename tasks to have test id and description
The tasks were previously renamed to just be the test id, due to
limitations in the custom_logger callbaack plugin. This was addressed,
and the test_report_result reports in the correct format now, so the
names can be reverted.
This makes reports more grokable, since the summary_result.log contains
a list of the passed/failed tests with the task names.