Add container tests to logging job (#219)
* [ci] Add container tests to the logging job

The test check for the containers that should exist on the compute node.

Co-authored-by: "Alex Yefimov <ayefimov@redhat.com>"

* [common] Update container tests

Add an additional fail condition and make the fail_msg more verbose
Update the container tests to use elevated permissions

* [ci][common] Use separate play for tests requiring elevated priviledges

The container tests need elevated permissions to run on the compute nodes.
This is not always going to be the case, so become is added to a separate play to only elevate priviliges for some tests.