[telemetry_logging] Fix failure condition in journalctl
failed_when and assert use `and` for for list items, and only
fails/passes when all conditions are met.

Currently, failed_when is used to trigger a test failure in the journal
tests, however, it is implemented incorrectly, since a list is used,
and the task will only fail if all conditions are met. This is
incorrect, the test should fail if any of the failure conditions are
met.

Using assert instead of failed_when means the test will fail if any of
the conditions are not met, which is the expected and correct behaviour.

This change switches to using assert, and check for the complementary
conditions.
Using assert also allows for some verbose fail messages to aid in
debugging.