[ci] Try to combine the FVT and tempest
This approach runs the FVTs before tempets and then collects the results after running test-operator
This means that the tempest tests will trigger a failure if they fail, and the FVTs will trigger a
failure only after the tempest tests are run, which means that the results are reported as expected.

This will work for github-check pipeline and the component pipeline.
However, this approach may not be usable with the unified jobs.