Set default values for *_test_id vars (#255)
The *test_id var was set for the common and telemetry_logging
roles previously, so that the tests would be reported as
passed/failed by custom_logger

Since report_result was updated to use XML files to determine
failure, the test ID does not need to be set.
This change adds a default value for these vars, so that we can start
removing the vars from test configs.