[ci] Copy results files immediately after running FVTs (#268)
If the tempest tests fail, the job ends.
In order to make sure the FVT results are available, they need to be
copied to their final destination before tempest tests are run.

To do this, a bool was added to the ci/report_results.yml playbook,
which disables reporting failures (the files get copied, but the results
are not checked for failures, which would end the job if there were
failures, therefor skipping tempest tests).

The report_results is run again after the tempest tests, but this time
the failures get reported.