[ci] Use pre_tests hooks for running FVTs (#272)
In some cases we were using post_deploy, and others pre_test for running the FVTs.
These happen sequentially, but for updates jobs, the post_deploy hooks are not run.
In order for the tests to be run in updates jobs, the pre_tests hook should be used.

JIRA: OSPRH-17110