[roles/common] allow using a dict to configure pod tests
The pod tests are currently configured by setting a
common_pod_list var that specifies the names of the pods.
These pods must all be in the same namespace and share the
same status_str and test_id. This means that the user needs
to run the role multiple times in order to check a list of
pods that are in different namespaces or have different
status_str values.

This change allow the configuration of the tests to either
remain the same or to use a dictionary of values per pod,
allowing to check for pods across multiple namespaces or
with multiple statuses and run the role once.