Add periodic job with volume pool metrics tests
To check volume pool metrics, we need a correct environment
configured with a storage backend - in this case ceph.
The job takes abet 20 minutes longer than our usual jobs,
so it was decided to run it as part of the periodic pipeline
instead of running it with each PR in the check pipeline.