[test_autoscaling] Remove the 5 minute pause before checking alarms
The tasks that follows does a repeated check for the expected results.
The 5 minute pause was rolled into the following task, by increasing the number of retries.
Adding in the 5 minutes into this task should mitigate the risk of not having enough time for
the alarm to effect to be triggered.