[OSPRH-21020] Adjust rabbitmq tests
A bugfix proposed in https://github.com/openstack-k8s-operators/telemetry-operator/pull/780
changes how rabbitmq metrics are being scraped. The FVT needs
to be adjusted for the tests to pass. For a limited time, both
versions of telemetry-operator (pre-bugfix and post-bugfix) are allowed
to pass the tests. A PR to remove the pre-bugfix tasks will be created in the
future.