Revert modify_openstackversions hook post_ctlplane
The https://github.com/infrawatch/feature-verification-tests/pull/315
PR modified the hook to be executed as pre_tests. This caused the
hook to be executed after our functional tests instead of before,
which caused failures.

Also this wasn't noticed, because of the "files" field we had defined
in zuul.yaml for the functional-tests-osp18, which prevented the
affected job to be executed on the original PR making everyone think
it passed successfuly.

This PR removes the "files" list from functional-tests-osp18 to
have the functional-tests-osp18 executed for this change and to prevent
similar mistakes in the future. The "irrelevant-files" list we have
defined at the job definition seems detailed enough to stop the job
from executing where we don't want it to.