[ci] Add explicit namespace to patch_observabilityclient (#321)
The tests don't explicitly log into a specific namespace.
When running the task that update the observabilityclient, the namespace
is omitted, so whatever namespace was last activated in openshift will
be used.
This commit updates the commands to explicitly state the namespace,
preventing errors when trying to connect to a non-existant pod.